refactor: combine app-core and app packages
