+++ packages/app-admin/src/base/ui/Logo.tsx
@@ -1,5 +1,5 @@
 import React, { useMemo } from "react";
-import { Compose, makeComposable } from "@webiny/app-core";
+import { Compose, makeComposable } from "@webiny/app";
 
 export const Logo = makeComposable("Logo", () => {
     return <LogoRenderer />;
