+++ packages/app-admin/src/base/ui/NotFound.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import { makeComposable } from "@webiny/app-core";
+import { makeComposable } from "@webiny/app";
 
 export const NotFound = makeComposable("NotFound", () => {
     return <NotFoundRenderer />;
