+++ packages/app-admin/tsconfig.json
@@ -3,7 +3,6 @@
   "include": ["src", "__tests__/**/*.ts"],
   "references": [
     { "path": "../app" },
-    { "path": "../app-core" },
     { "path": "../app-security" },
     { "path": "../app-wcp" },
     { "path": "../form" },
@@ -23,8 +22,6 @@
       "~tests/*": ["./__tests__/*"],
       "@webiny/app/*": ["../app/src/*"],
       "@webiny/app": ["../app/src"],
-      "@webiny/app-core/*": ["../app-core/src/*"],
-      "@webiny/app-core": ["../app-core/src"],
       "@webiny/app-security/*": ["../app-security/src/*"],
       "@webiny/app-security": ["../app-security/src"],
       "@webiny/app-wcp/*": ["../app-wcp/src/*"],
