--- packages/app-core/.babelrc.js
@@ -1,1 +1,0 @@
-module.exports = require("@webiny/project-utils").createBabelConfigForReact({ path: __dirname });
