--- packages/app-core/README.md
@@ -1,7 +1,0 @@
-# @webiny/app-core
-[![](https://img.shields.io/npm/dw/@webiny/app-core.svg)](https://www.npmjs.com/package/@webiny/app-core) 
-[![](https://img.shields.io/npm/v/@webiny/app-core.svg)](https://www.npmjs.com/package/@webiny/app-core)
-[![code style: prettier](https://img.shields.io/badge/code_style-prettier-ff69b4.svg?style=flat-square)](https://github.com/prettier/prettier)
-[![PRs Welcome](https://img.shields.io/badge/PRs-welcome-brightgreen.svg?style=flat-square)](http://makeapullrequest.com)
-
-Webiny admin app core.
