--- packages/app-core/package.json
@@ -1,63 +1,0 @@
-{
-  "name": "@webiny/app-core",
-  "version": "0.0.0",
-  "main": "index.js",
-  "description": "Core framework of the Webiny React apps.",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/webiny/webiny-js.git"
-  },
-  "author": "Pavel Denisjuk",
-  "license": "MIT",
-  "dependencies": {
-    "@babel/runtime": "^7.19.0",
-    "@types/react": "17.0.39",
-    "@webiny/app": "0.0.0",
-    "@webiny/plugins": "0.0.0",
-    "@webiny/react-composition": "0.0.0",
-    "@webiny/react-router": "0.0.0",
-    "lodash.debounce": "^4.0.8",
-    "react": "17.0.2",
-    "react-dom": "17.0.2"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.19.3",
-    "@babel/core": "^7.19.3",
-    "@babel/plugin-proposal-class-properties": "^7.18.6",
-    "@babel/preset-env": "^7.19.4",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.18.6",
-    "@types/lodash.debounce": "^4.0.6",
-    "@webiny/cli": "0.0.0",
-    "@webiny/project-utils": "0.0.0",
-    "babel-plugin-named-asset-import": "^1.0.0-next.3e165448",
-    "rimraf": "^3.0.2",
-    "ttypescript": "^1.5.12",
-    "typescript": "4.7.4"
-  },
-  "publishConfig": {
-    "access": "public",
-    "directory": "dist"
-  },
-  "scripts": {
-    "build": "yarn webiny run build",
-    "watch": "yarn webiny run watch"
-  },
-  "svgo": {
-    "plugins": {
-      "removeViewBox": false
-    }
-  },
-  "adio": {
-    "ignore": {
-      "dependencies": [
-        "@svgr/webpack",
-        "@emotion/core",
-        "@types/mime"
-      ],
-      "peerDependencies": [
-        "react-dom"
-      ]
-    }
-  }
-}
