--- packages/app-core/src/index.ts
@@ -1,13 +1,0 @@
-// Composition - we re-export this for ease of use
-export * from "@webiny/react-composition";
-export type { HigherOrderComponent, ComposeProps, ComposableFC } from "@webiny/react-composition";
-
-// App framework
-export * from "./App";
-export type { AppProps } from "./App";
-
-export * from "./core/Plugins";
-export * from "./core/Provider";
-export * from "./core/Routes";
-export * from "./utils/DebounceRender";
-export * from "./utils/createProviderPlugin";
