--- packages/app-core/tsconfig.build.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "include": ["src"],
-  "references": [
-    { "path": "../app/tsconfig.build.json" },
-    { "path": "../plugins/tsconfig.build.json" },
-    { "path": "../react-composition/tsconfig.build.json" },
-    { "path": "../react-router/tsconfig.build.json" }
-  ],
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": { "~/*": ["./src/*"], "~tests/*": ["./__tests__/*"] },
-    "baseUrl": "."
-  }
-}
