--- packages/app-core/tsconfig.json
@@ -1,28 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "include": ["src", "__tests__/**/*.ts"],
-  "references": [
-    { "path": "../app" },
-    { "path": "../plugins" },
-    { "path": "../react-composition" },
-    { "path": "../react-router" }
-  ],
-  "compilerOptions": {
-    "rootDirs": ["./src", "./__tests__"],
-    "outDir": "./dist",
-    "declarationDir": "./dist",
-    "paths": {
-      "~/*": ["./src/*"],
-      "~tests/*": ["./__tests__/*"],
-      "@webiny/app/*": ["../app/src/*"],
-      "@webiny/app": ["../app/src"],
-      "@webiny/plugins/*": ["../plugins/src/*"],
-      "@webiny/plugins": ["../plugins/src"],
-      "@webiny/react-composition/*": ["../react-composition/src/*"],
-      "@webiny/react-composition": ["../react-composition/src"],
-      "@webiny/react-router/*": ["../react-router/src/*"],
-      "@webiny/react-router": ["../react-router/src"]
-    },
-    "baseUrl": "."
-  }
-}
