--- packages/app-core/webiny.config.js
@@ -1,8 +1,0 @@
-const { createWatchPackage, createBuildPackage } = require("@webiny/project-utils");
-
-module.exports = {
-    commands: {
-        build: createBuildPackage({ cwd: __dirname }),
-        watch: createWatchPackage({ cwd: __dirname })
-    }
-};
