+++ packages/app-security/tsconfig.build.json
@@ -3,7 +3,6 @@
   "include": ["src"],
   "references": [
     { "path": "../app/tsconfig.build.json" },
-    { "path": "../app-core/tsconfig.build.json" },
     { "path": "../plugins/tsconfig.build.json" }
   ],
   "compilerOptions": {
