+++ packages/app/package.json
@@ -22,6 +22,7 @@
     "@webiny/i18n": "0.0.0",
     "@webiny/i18n-react": "0.0.0",
     "@webiny/plugins": "0.0.0",
+    "@webiny/react-composition": "0.0.0",
     "@webiny/react-router": "0.0.0",
     "@webiny/ui": "0.0.0",
     "apollo-cache": "^1.3.5",
@@ -36,6 +37,7 @@
     "graphql": "^15.7.2",
     "invariant": "^2.2.4",
     "lodash": "^4.17.4",
+    "lodash.debounce": "^4.0.8",
     "nanoid": "^3.1.30",
     "react": "17.0.2",
     "react-dom": "17.0.2",
@@ -48,6 +50,7 @@
     "@babel/preset-env": "^7.19.4",
     "@babel/preset-react": "^7.18.6",
     "@babel/preset-typescript": "^7.18.6",
+    "@types/lodash.debounce": "^4.0.7",
     "@types/warning": "^3.0.0",
     "@webiny/cli": "0.0.0",
     "@webiny/project-utils": "0.0.0",
