+++ packages/app/src/core/AddRoute.tsx
@@ -1,5 +1,5 @@
 import React, { useEffect } from "react";
-import { useApp } from "~/index";
+import { useApp } from "~/App";
 import { Route, RouteProps } from "@webiny/react-router";
 
 export const AddRoute: React.FC<RouteProps> = props => {
