+++ packages/app/src/core/Plugins.tsx
