+++ packages/app/src/index.ts
@@ -2,3 +2,16 @@ export { AddQuerySelectionPlugin } from "./plugins/AddQuerySelectionPlugin";
 export { ApolloLinkPlugin } from "./plugins/ApolloLinkPlugin";
 export { RoutePlugin } from "./plugins/RoutePlugin";
 export { ApolloCacheObjectIdPlugin } from "./plugins/ApolloCacheObjectIdPlugin";
+
+// Composition - we re-export this for ease of use
+export * from "@webiny/react-composition";
+export type { HigherOrderComponent, ComposeProps, ComposableFC } from "@webiny/react-composition";
+
+// App framework
+export * from "./App";
+export type { AppProps } from "./App";
+export * from "./core/Plugins";
+export * from "./core/Provider";
+export * from "./core/AddRoute";
+export * from "./core/DebounceRender";
+export * from "./core/createProviderPlugin";
